+++ src/eval.c
@@ -22,7 +22,8 @@
 #include "bits.h"
 #include "board.h"
 #include "endgame.h"
-#include "nn.h"
+#include "nn/accumulator.h"
+#include "nn/eval.h"
 #include "search.h"
 #include "uci.h"
 #include "util.h"
