+++ src/uci.c
@@ -29,7 +29,8 @@
 #include "move.h"
 #include "movegen.h"
 #include "movepick.h"
-#include "nn.h"
+#include "nn/accumulator.h"
+#include "nn/eval.h"
 #include "perft.h"
 #include "pyrrhic/tbprobe.h"
 #include "search.h"
